requirements.txt
pytorch==1.7.1
pillow
pandas
